MarkovBounds.jl Documentation

Documentation for MarkovBounds.

Bounds on Stationary Moments of Markov Processes

MarkovBounds.stationary_polynomial — Method

stationary_polynomial(MP::MarkovProcess, v::APL, d::Int, solver)

returns a lower bound on the expecation of a polynomial observables v(x) at steady state of the Markov process MP. The bound is computed based on an SOS program over a polynomial of degree at most d; the bounds can be tightened by increasing d. The program is solved with solver.

MarkovBounds.stationary_mean — Method

stationary_mean(MP::MarkovProcess, v::APL, d::Int, solver)

returns lower and upper bound on the observable v(x) at steady state of the Markov process MP. Bounds are computed based on SOS programs over a polynomial of degree at most d; the bounds can be tightened by increasing d. The program is solved with solver.

MarkovBounds.stationary_mean — Function

stationary_mean(rn::ReactionSystem, S0::Dict, S, d::Int, solver,
		scales = Dict(s => 1 for s in species(rn));
		auto_scaling = false)

returns lower and upper bound on the mean of species S of the reaction network rn with initial condition S0 (for all species!). The bound is based on an SOS program of order d solved via solver; the bounds can be tightened by increasing d.

For numerical stability, it is recommended to provide scales of the expected magnitude of molecular counts for the different species at steady state. If the system is closed it is also possible to enable auto_scaling which will find the maximum molecular counts for each species under stoichiometry constraints (via LP).

If the initial condition of the reaction network under investigation is unknown or irrelevant, simply call

stationary_mean(rn::ReactionSystem, S, d::Int, solver,
		scales = Dict(s => 1 for s in species(rn))).

MarkovBounds.stationary_variance — Method

stationary_variance(MP::MarkovProcess, v::APL, d::Int, solver)

returns SOS program of degree d for computation of an upper bound on the variance of a polynomial observables v at steady state of the Markov process MP.

MarkovBounds.stationary_variance — Function

stationary_variance(rn::ReactionSystem, S0, x, d::Int, solver,
		    scales = Dict(s => 1 for s in species(rn));
		    auto_scaling = false)

returns upper bound on the variance of species S of the reaction network rn with initial condition S0 (for all species!). The bound is based on an SOS program of degree d solved via solver; the bound can be tightened by increasing d.

For numerical stability, it is recommended to provide scales of the expected magnitude of molecular counts for the different species at steady state. If the system is closed it is also possible to enable auto_scaling which will find the maximum molecular counts for each species under stoichiometry constraints (via LP).

If the initial condition of the reaction network under investigation is unknown or irrelevant, simply call

stationary_variance(rn::ReactionSystem, S, d::Int, solver,
		    scales = Dict(s => 1 for s in species(rn)))

MarkovBounds.stationary_covariance_ellipsoid — Method

stationary_covariance_ellipsoid(MP::MarkovProcess, v::Vector{<:APL}, d::Int, solver)

returns an upper on the volume of the covariance ellipsoid of a vector of polynomial observables v(x), i.e., det(𝔼(v(x)v(x)ᵀ)), at steady state of the Markov process MP. The bounds are computed via an SOS program of degree d, hence can be tightened by increasing d. This computation requires a solver that can deal with exponential cone constraints.

MarkovBounds.stationary_covariance_ellipsoid — Function

stationary_covariance_ellipsoid(rn::ReactionSystem, S0::Dict, S::AbstractVector, d::Int, solver,
				scales = Dict(s => 1 for s in species(rn));
				auto_scaling = false)

returns an upper on the volume of the covariance ellipsoid of any subset S of the chemical species in the reaction network rn, i.e., det(𝔼(SSᵀ)), at steady state of the associated jump process. The reaction network is assumed to have the deterministic initial state S0 (all species must be included here!). The bounds are computed via an SOS program of degree d, hence can be tightened by increasing d. This computation requires a solver that can deal with exponential cone constraints.

For numerical stability, it is recommended to provide scales of the expected magnitude of molecular counts for the different species at steady state. If the system is closed it is also possible to enable auto_scaling which will find the maximum molecular counts for each species under stoichiometry constraints (via LP).

If the initial condition of the reaction network under investigation is unknown or irrelevant, simply call

stationary_covariance_ellipsoid(rn::ReactionSystem, S, d::Int, solver,
				scales = Dict(s => 1 for s in species(rn)))

Bounds on Transient Moments of Markov Processes

MarkovBounds.transient_polynomial — Method

transient_polynomial(MP::MarkovProcess, μ0::Dict, v::APL, d::Int, trange::AbstractVector{<:Real}, solver)

returns a lower bound on 𝔼[v(x(T))] where v is a polynomial and x(T) the state of the Markov process MP at time T. μ0 encodes the distribution of the initial state of the process in terms of its moments; specifically, it maps monomials to the respective moments of the initial distribution. trange is an ordered collection of time points used to discretize the time horizon [0,T], i.e., trange[end] = T. Populating trange and increasing d improves the computed bound.

MarkovBounds.transient_mean — Method

transient_mean(MP::MarkovProcess, μ0::Dict, x::APL, d::Int, trange::AbstractVector{<:Real}, solver)

returns a lower and upper bound on 𝔼[v(x(T))] where v is a polynomial and x(T) the state of the Markov process MP at time T. μ0 encodes the distribution of the initial state of the process in terms of its moments; specifically, it maps monomials to the respective moments of the initial distribution. trange is an ordered collection of time points used to discretize the time horizon [0,T], i.e., trange[end] = T. Populating trange and increasing d improves the computed bounds.

MarkovBounds.transient_mean — Function

transient_mean(rn::ReactionSystem, S0::Dict, S, d::Int, trange::AbstractVector{<:Number}, solver,
		scales = Dict(s => 1 for s in species(rn));
		auto_scaling = false)

returns a lower and upper bound on the mean of the molecular count of species S in reaction network rn at time T. S0 refers to the deterministic initial state of the reaction system (including all species!). trange is an ordered collection of time points used to discretize the time horizon [0,T], i.e., trange[end] = T. Populating trange and increasing d improves the computed bounds.

For numerical stability, it is recommended to provide scales of the expected magnitude of molecular counts for the different species at steady state. If the system is closed it is also possible to enable auto_scaling which will find the maximum molecular counts for each species under stoichiometry constraints (via LP).

MarkovBounds.transient_variance — Method

transient_variance(MP::MarkovProcess, μ0::Dict, v::APL, d::Int, trange::AbstractVector{<:Real}, solver)

returns an upper bound on 𝔼[v(x(T))²] - 𝔼[v(x(T))]² where v is a polynomial and x(T) the state of the Markov process MP at time T.

MarkovBounds.transient_variance — Function

transient_variance(rn::ReactionSystem, S0::Dict, S, d::Int, trange::AbstractVector{<:Real}, solver,
		    scales = Dict(s => 1 for s in species(rn));
			auto_scaling = false)

returns an upper bound on the variance of species S in the reaction network rn at time T. S0 refers to the deterministic initial state of the reaction system (including all species!). trange is an ordered collection of time points used to discretize the time horizon [0,T], i.e., trange[end] = T. Populating trange and increasing d improves the computed bounds.

For numerical stability, it is recommended to provide scales of the expected magnitude of molecular counts for the different species at steady state. If the system is closed it is also possible to enable auto_scaling which will find the maximum molecular counts for each species under stoichiometry constraints (via LP).

MarkovBounds.transient_covariance_ellipsoid — Method

transient_covariance_ellipsoid(MP::MarkovProcess, μ0::Dict, v::Vector{APL}, d::Int, trange::AbstractVector{<:Real}, solver)

returns an upper bound on the volume of the covariance ellipsoid det(𝔼(v(x(T))v(x(T))ᵀ)), where v is a polynomial and x(T) the state of the Markov process MP at time T.

MarkovBounds.transient_covariance_ellipsoid — Function

transient_covariance_ellipsoid(rn::ReactionSystem, S0::Dict, S::AbstractVector, d::Int, trange::AbstractVector{<:Real}, solver,
						scales = Dict(s => 1 for s in species(rn));
						auto_scaling = false)

returns an upper bound on the volume of the covariance ellipsoid associated with any collection of chemical species in the reaction network rn at time T. S0 refers to the deterministic initial state of the reaction system (including all species!). trange is an ordered collection of time points used to discretize the time horizon [0,T], i.e., trange[end] = T. Populating trange and increasing d improves the computed bounds.

For numerical stability, it is recommended to provide scales of the expected magnitude of molecular counts for the different species at steady state. If the system is closed it is also possible to enable auto_scaling which will find the maximum molecular counts for each species under stoichiometry constraints (via LP).

Bounds on Stochastic Optimal Control Problems

MarkovBounds.optimal_control — Method

optimal_control(CP::ControlProcess, μ0::Dict, d::Int, trange::AbstractVector{<:Real}, solver)

returns a lower bound on the objective value of the (stochastic) optimal control problem specified by CP. μ0 encodes information about the distribution of the initial state of the process; specifically, μ0 maps a given monomial to the corresponding moment of the initial distribution. trange refers to an ordered set of time points discretizing the control horizon. trange[end] should coincide with the end of the control horizon, i.e., trange[end] = Inf in case of an infinite horizon problem. The bound is computed via a SOS program of degree d solved with an appropriate method given by solver.

The bound can be tightened by populating trange or increasing d.